+++ flight-service/src/main/java/ru/shemich/flightservice/repository/AirportRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.shemich.flightservice.model.Airport;
 
+import java.util.Optional;
+
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
+public interface AirportRepository extends JpaRepository<Airport, Long> {
+
 }
